requirements.txt
blinker==1.7.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cryptography==42.0.5
Flask==3.0.3
google-genai>=1.0.0
gunicorn==22.0.0
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.3
joblib==1.4.0
MarkupSafe==2.1.5
msgspec==0.18.6
numpy==1.26.4
packaging==24.0
pandas==2.2.2
pycparser==2.22
PyMuPDF==1.24.14
python-dateutil==2.9.0.post0
pytz==2024.1
regex==2024.4.16
requests==2.31.0
six==1.16.0
tqdm==4.66.2
sentence-transformers==5.2.2
tzdata==2024.1
urllib3==2.2.1
Werkzeug==3.0.2
XlsxWriter==3.2.0
pytest==8.0.0
pytest-cov==4.1.0